rename @latticexyz/config/node to @latticexyz/config/cli
