+++ packages/cli/src/commands/tsgen.ts
@@ -1,5 +1,5 @@
 import type { CommandModule } from "yargs";
-import { loadConfig } from "@latticexyz/config/node";
+import { loadConfig } from "@latticexyz/config/cli";
 import { StoreConfig } from "@latticexyz/store";
 import { tsgen } from "../render-ts/tsgen";
