+++ packages/config/src/cli/index.ts
