+++ packages/config/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/library/index.ts", "src/register/index.ts", "src/node/index.ts"],
+  entry: ["src/library/index.ts", "src/register/index.ts", "src/cli/index.ts"],
   target: "esnext",
   format: ["esm"],
   dts: false,
